rename javascript files to .js
