+++ src/utils/userNameGenerator.js
@@ -1,4 +1,4 @@
-import User from "../models/user.mjs";
+import User from "../models/user.js";
 
 export const generateUserName = async () => {
   const recentRegisteredUser = await User.find().sort({ _id: -1 }).limit(1);
